requirements.txt
# Backend dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2

# Frontend dependencies
streamlit==1.28.1

# Agent system dependencies
crewai==0.28.0
langchain==0.0.335
langchain-community==0.0.13

# Utilities
python-dotenv==1.0.0
google-search-results==2.4.2
openai==1.3.5
requests==2.31.0

# Additional dependencies
pydantic-settings>=2.1.0
litellm>=1.16.0
pytest>=8.0.0
httpx>=0.26.0
python-jose[cryptography]>=3.3.0
beautifulsoup4>=4.12.0
markdown>=3.5.0 